+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.1]
+
+### Added
+
+- Parsing 'Randomness' values
+
+### Changed
+
+- Fixed `catch2TestExplorer.defaultRngSeed`
+
 ## [1.1.0]
 
 ### Added